Unconditionally send commands
We were being smart and assumed that all commands set K4 state.
This is true for most commands, but not all. In particular, the SW
commands modify the K4 state, but they don't directly represent state.
Now, we'll send commands even if we recently sent the same command.
In the future, we could distinguish between those commands that
represent state and those that don't, but it likely isnt worth it.