Workaround a K4 bug dealing with 100 KHz vfo steps
At least on the current version Production V36 and Beta V37.1,
On startup, regardless of the VFO A step size, it reports a step
size for VFO A of 100 KHz. The step size for VFO B is reported
correctly.

Also, if you try to select 100 KHz steps, it shows 100 Hz steps but
reports via the API that it has 100 KHz steps.

We disallow seting 100 KHz step size via the API and if we receive
a step size of 100 KHz, we force it to be 100 Hz, matching what is
on the K4 display.