Handle a partial cmd at the end of a read buffer
When a buffer is received that does not end in ';', a partial command is
at the end of the buffer, and must be prepended to the following buffer that
is received. The logic handling this was wrong. I hope it is correct now.

Thanks to Dick, N4UN, for reporting the bug.